refactor: test reacthookform on radio
