+++ libs/stack-ui/src/components/fields/Radio/Radio.interface.ts
@@ -1,8 +1,10 @@
 import type React from 'react'
 import type { Noop, RefCallBack } from 'react-hook-form'
-import type { TDefaultComponent } from '../../../types/components'
+import type { TDefaultComponent, TReactHookForm } from '../../../types/components'
 
-export interface TRadioProps extends TDefaultComponent {
+export type TFieldReactHookForm = TReactHookForm & TDefaultComponent
+
+export interface TRadioProps extends TFieldReactHookForm {
   id: string
   label: string
   name: string
@@ -12,7 +14,6 @@ export interface TRadioProps extends TDefaultComponent {
   errorMessage?: string
   ariaLabel?: string
   value: string
-  onBlur?: Noop
   onChange?: (isSelected: boolean) => void
   inputRef?: RefCallBack
   icon?: React.ReactNode
