+++ libs/stack-ui/src/components/fields/Radio/index.tsx
@@ -1,14 +1,17 @@
 'use client'
 
+import { isEmpty } from 'lodash'
 import { useRef } from 'react'
 import { useRadio, VisuallyHidden, useFocusRing, FocusRing, mergeProps } from 'react-aria'
+import { get, useFormContext } from 'react-hook-form'
+import { useTranslation } from 'react-i18next'
 import { useRadioGroupCtx } from '../../../providers/RadioGroup/RadioGroup'
 import useThemeContext from '../../../providers/Theme/hooks'
 import Typography from '../../Typography'
 import type { TRadioProps } from './Radio.interface'
 
 const Radio = (props: TRadioProps) => {
-  const { id, label, isDisabled = false, ariaLabel, themeName = 'radio', tokens, customTheme, value } = props
+  const { id, label, themeName = 'radio', tokens, customTheme, value } = props
   const state = useRadioGroupCtx()
   const ref = useRef<HTMLInputElement | null>(null)
   // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
@@ -16,7 +19,7 @@ const Radio = (props: TRadioProps) => {
   const isSelected = state?.selectedValue === value
   const { focusProps } = useFocusRing()
 
-  const radioTokens = { ...tokens, selected: isSelected, isDisabled }
+  const radioTokens = { ...tokens, selected: isSelected }
   const labelTheme = useThemeContext(`${themeName}.label`, radioTokens, customTheme)
   const radioTheme = useThemeContext(`${themeName}.radio`, radioTokens, customTheme)
   const selectedMarkTheme = useThemeContext(`${themeName}.selectedMark`, radioTokens, customTheme)
@@ -24,7 +27,7 @@ const Radio = (props: TRadioProps) => {
   return (
     <div>
       <FocusRing focusRingClass="has-focus-ring" within>
-        <label htmlFor={id} className={labelTheme} aria-label={ariaLabel} aria-disabled={isDisabled}>
+        <label htmlFor={id} className={labelTheme}>
           <VisuallyHidden>
             <input type="radio" {...mergeProps(focusProps, inputProps)} ref={ref} />
           </VisuallyHidden>
@@ -36,5 +39,18 @@ const Radio = (props: TRadioProps) => {
       </FocusRing>
     </div>
   )
+}
+
+export const ReactHookFormRadio = (props: TRadioProps) => {
+  const { name, required } = props
+  const { register, formState } = useFormContext()
+  const error: Error = get(formState.errors, name)
+  const errMsg = error?.message ?? undefined
+  const { t } = useTranslation('components')
+  const { ref: refCallback, ...rest } = register(name, {
+    required: required ? t('FORM.ERROR.REQUIRED') ?? 'required' : false,
+  })
+
+  return <Radio inputRef={refCallback} {...rest} {...props} isError={!isEmpty(errMsg)} errorMessage={errMsg} />
 }
 export default Radio
