+++ libs/stack-ui/src/components/fields/RadioGroup/RadioGroup.interface.ts
@@ -1,16 +1,10 @@
 import type React from 'react'
-import type { RefCallBack } from 'react-hook-form'
-import type { TDefaultComponent, TReactHookForm } from '../../../types/components'
+import type { TDefaultComponent } from '../../../types/components'
 
-export type TFieldReactHookForm = TReactHookForm & TDefaultComponent
-
-export interface TRadioGroupProps extends TFieldReactHookForm {
+export interface TRadioGroupProps extends TDefaultComponent {
   children: React.ReactNode
   isError?: boolean
   isDisabled?: boolean
   errorMessage?: string
   label?: string
-  fieldRef?: RefCallBack
-  required?: boolean
-  name: string
 }
